Move SmartContractBlockHeader to Core
