+++ src/Stratis.Bitcoin.Features.SmartContracts/PoA/SmartContractPoABlockHeader.cs
@@ -1,6 +1,7 @@
 ﻿using NBitcoin;
 using Stratis.Bitcoin.Features.PoA;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Interfaces;
 using TracerAttributes;
 using uint256 = NBitcoin.uint256;
