+++ src/Stratis.Bitcoin.Features.SmartContracts/ReflectionExecutor/Controllers/ReceiptSearcher.cs
@@ -5,6 +5,7 @@ using NBitcoin;
 using Stratis.Bitcoin.Interfaces;
 using Stratis.SmartContracts.CLR;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Interfaces;
 using Stratis.SmartContracts.Core.Receipts;
 
 namespace Stratis.Bitcoin.Features.SmartContracts.ReflectionExecutor.Controllers
