+++ src/Stratis.Bitcoin.Features.SmartContracts/SmartContractBlockHeader.cs
@@ -1,5 +1,6 @@
 ﻿using NBitcoin;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Interfaces;
 
 namespace Stratis.Bitcoin.Features.SmartContracts
 {
