+++ src/Stratis.SmartContracts.IntegrationTests/ContractInternalTransferTests.cs
@@ -8,6 +8,7 @@ using Stratis.SmartContracts.CLR.Compilation;
 using Stratis.SmartContracts.CLR.Local;
 using Stratis.SmartContracts.CLR.Serialization;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Interfaces;
 using Stratis.SmartContracts.Core.Util;
 using Stratis.SmartContracts.Tests.Common.MockChain;
 using Xunit;
