+++ src/Stratis.SmartContracts.IntegrationTests/ContractParameterSerializationTests.cs
@@ -7,6 +7,7 @@ using Stratis.SmartContracts.CLR;
 using Stratis.SmartContracts.CLR.Compilation;
 using Stratis.SmartContracts.CLR.Serialization;
 using Stratis.SmartContracts.Core;
+using Stratis.SmartContracts.Core.Interfaces;
 using Stratis.SmartContracts.Tests.Common.MockChain;
 using Xunit;
